Add role authorization middleware and reuse for admin routes
